add exception

+++ src/main/java/io/javabrains/services/AdviserServices.java
@@ -1,12 +1,10 @@
 package io.javabrains.services;
 import java.util.List;
 import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import io.javabrains.dao.AdviserRepo;
-import io.javabrains.dao.SectionRepository;
-import io.javabrains.exceptionHandler.AdviserExceptionHandler;
+import io.javabrains.exception.AdviserException;
 import io.javabrains.model.Adviser;
 import io.javabrains.model.Section;
 
@@ -22,37 +20,46 @@ public class AdviserServices {
 	}
 	
 	public Adviser getAdviserById(int adviserId) {
-//		Optional<Adviser> adviserID = adviserRepo.findById(adviserId);
-//		if(!adviserID.isPresent()) {
-//			throw new AdviserExceptionHandler("AdviserID: "+adviserId+" does not exist !!!");
-//		}
+		Optional<Adviser> adviserID = adviserRepo.findById(adviserId);
+		if(!adviserID.isPresent()) {
+			throw new AdviserException("AdviserID: "+adviserId+" does not exist !!!");
+		}
 			return adviserRepo.findById(adviserId).get();	
 	}
 	
 	public void addAdviser(Adviser adviser) {
-//		int adviserId = adviser.getAdviserId();
-//		Optional<Adviser> adviserID = adviserRepo.findById(adviserId);
-//		if(adviserID.isPresent()) {
-//			throw new AdviserExceptionHandler("AdviserID: "+adviserId+" already exist !!!");
-//		}
+		int adviserId = adviser.getAdviserId();
+		Optional<Adviser> adviserID = adviserRepo.findById(adviserId);
+		if(adviserID.isPresent()) {
+			throw new AdviserException("AdviserID: "+adviserId+" already exist !!!");
+		}
 		adviserRepo.save(adviser);
 	}
 	
 	public void updateAdviser(int adviserId, Adviser adviser) {
-//		Optional<Adviser> adviserID = adviserRepo.findById(adviserId);
-//		if(!adviserID.isPresent()) {
-//			throw new AdviserExceptionHandler("AdviserID: "+adviserId+" does not exist");
-//		}
-//		adviser.setSection(sections);
-//		sections.setAdviser(adviser);
+		Optional<Adviser> adviserID = adviserRepo.findById(adviserId);
+		if(!adviserID.isPresent()) {
+			throw new AdviserException("AdviserID: "+adviserId+" does not exist");
+		}
+		//adviser.setSection(sections);
+		//sections.setAdviser(adviser);
+		Adviser existingAdviser = adviserRepo.findById(adviserId).get();
+		Section adviserSec = existingAdviser.getSection();
+		adviser.setSection(adviserSec);
 		adviserRepo.save(adviser);
 	}
 	
 	public void deleteAdviser(int adviserId) {
-//		Optional<Adviser> adviserID = adviserRepo.findById(adviserId);
-//		if(!adviserID.isPresent()) {
-//			throw new AdviserExceptionHandler("AdviserID: "+adviserId+" does not exist");
-//		}
+		Optional<Adviser> adviser = adviserRepo.findById(adviserId);
+		if(!adviser.isPresent()) {
+			throw new AdviserException("AdviserID: "+adviserId+" does not exist");
+		}
+		//Adviser existingAdviser = adviserRepo.findById(adviserId).get();
+		Section adviserSection = adviser.get().getSection();
+		if(adviserSection != null) {
+			adviserSection.setAdviser(null);
+		}
 		adviserRepo.deleteById(adviserId);
+		
 	}
 }
