+++ src/main/java/io/javabrains/services/StudentServices.java
@@ -1,10 +1,13 @@
 package io.javabrains.services;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import io.javabrains.dao.StudentRepository;
+import io.javabrains.exception.StudentException;
+import io.javabrains.model.Section;
 import io.javabrains.model.Student;
 
 @Service
@@ -21,18 +24,43 @@ public class StudentServices {
 	}
 	
 	public Student getStudentById(int studentId) {
+		Optional<Student> studentID = studentRepo.findById(studentId);
+		if(!studentID.isPresent()) {
+			throw new StudentException("StudentID: "+studentId+" does not exist!!!");
+		}
 		return studentRepo.findById(studentId).get();
 	}
 	
 	public void addStudent(Student student) {
+		int findStudentId = student.getStudentId();
+		Optional<Student> studentID = studentRepo.findById(findStudentId);
+		if(studentID.isPresent()) {
+			throw new StudentException("StudentID: "+findStudentId+" already exist");
+		}
 		studentRepo.save(student);
 	}
 	
 	public void updateStudent(int studentId,Student student) {
+		Optional<Student> studentID = studentRepo.findById(studentId);
+		if(!studentID.isPresent()) {
+			throw new StudentException("StudentID "+studentId+" does not exist");
+		}
+		Student existingStudent = studentRepo.findById(studentId).get();
+		Section studentSection = existingStudent.getSection();
+		student.setSection(studentSection);
 		studentRepo.save(student);
 	}
 	
 	public void deleteAdviser(int studentId) {
+		Optional<Student> studentID = studentRepo.findById(studentId);
+		if(!studentID.isPresent()) {
+			throw new StudentException("StudentID "+studentId+" does not exist");
+		}
+		Student existingStudent = studentRepo.findById(studentId).get();
+		Section studentSection = existingStudent.getSection();
+		if(studentSection != null) {
+			studentSection.setStudents(null);
+		}
 		studentRepo.deleteById(studentId);
 	}
 }
